Move enum SortOder to API
