+++ backend/eln/eln-api/src/main/java/com/epam/indigoeln/eln/model/SortOrder.java
@@ -1,0 +1,16 @@
+package com.epam.indigoeln.eln.model;
+
+import lombok.Getter;
+
+@Getter
+public enum SortOrder {
+    EARLIEST("Earliest"),
+    LATEST("Latest"),
+    DEFAULT("Default");
+
+    private final String value;
+
+    SortOrder(String value) {
+        this.value = value;
+    }
+}
