+++ backend/eln/eln-core/src/main/java/com/epam/indigoeln/eln/controller/ProjectResource.java
@@ -32,7 +32,7 @@ public class ProjectResource implements ProjectAPI {
     }
 
     @Override
-    public @NotNull @Valid Page<ProjectDTO> getProjects(@Nullable String search, @Nullable String sort, @Nullable Boolean createdByMe, @Valid Paging paging) {
+    public @NotNull @Valid Page<ProjectDTO> getProjects(@Nullable String search, @Nullable SortOrder sort, @Nullable Boolean createdByMe, @Valid Paging paging) {
         return projectService.getProjects(search, sort, createdByMe, paging);
     }
